Transferred much MCMC specific content from main routine to ptmcmc.
About 30% of the main routine has been moved to a new ptmcmc_sampler
class, derived from bayes_sampler.  This class uses a new Optioned
class in options.hh (moved to ptmcmc) which provides a general
interface for objects to define 'options' which may be command-line
flags in the main routine. There are now just a few hooks to
bayes_sampler in the main routine, and even fewer to ptmcmc_sampler
itself.

The hooks are: the ptmcmc_sampler constructor, a routine for
defining the sampler characteristics after options are processed
ptmcmc_sampler::setup(). A clone() routine, an initialization
routine where the mcmc chains and proposal objects are internally
instantiated, a run() routine, and an analyze routine.  The last
has a slightly messy interface because it involves dumping example
signals, which ptmcmc doesn't otheriwse know of.  This still has
a messy argument list which can probably be improved by adding more
structure to the new bayes_signal interface.

The code has been tested with the same test run of recent commits
showing agreement through multiple cycles of chain runs.