Adding in the script for glens testing from before. Run by e.g.: tests.py full_glens_test_set 8 path-to-gleam
This will generate a lot of output running gleam mag-map a lot of times, then a
summary indicating which of the test-set cases were consistent at the 1e-8 level
(hence the 8).

As is this doesn't work without compiling gleam_quad which is untested...
There may be other issues.  Use a much smaller test-set file for debugging.