Here is a commit of an update to the GLens::compute_trajectory code focusing mainly on the integration variant of the routine. Elements:  -Simplified the loop structure in that routine for clarity and robustness 	(verifed against Oct version of the code)  -Improved the tests.py script for better testing (see the help message)  -Fixed --precision flag for mag-map option  -Identified issues affecting the runs with the integration option     -Need to verify 3 or 5 images before switching to evolution     -Kappa coeff was an int, and therfore effectively zero     -Relaxed tolerance for root checking as some cases had many missed roots      Manyly only affects high-prec comparisons with quad-prec.  -Made the tuning parameters accessible from the command line
Following these changes, it seems taht the optimal tuning of kappa is near
the default of 0.1 and the int_tol can vary by a couple orders of mag from
the default of 1e-10, without damaging the results on a mini test set, which
included all those cases which didn't pass at 1e-7 level before making changes.
With these fixes and in this tuning range the runs pass the test at the 1% level
with one point that fails the 1e-3 test.