Changes in how time referencing is handled and an update in snapshot.py
First, in snapshot.py we have made small changes to support the new format
for providing the datafile in gleam (and the dropping of gle_ prefix in
output files.  When the original run uses the new format, ie -OGLE_data=xx
then only one argument (either .out or .dat file) is needed.

The time referencing changes are motivated by the need to complete the
abstraction of the Trajectory object in mlsignal.hh.  In the process
we have discovered that our prior handling of the tmax parameter was
(still) incorrect (depite fixes last summer).  The new tmax parameter
is in fact 1/2 of the time offset from the closet approach point to
the reference (nominally data peak) time.  A fix for this will be
applied in an upcoming commit. Here we implement the following changes:

 -mldata:organized the handling of time shifting into the base class
	By default the behavior is as before.  There is a new option
	to alter the time referencing by calling set_reference_time(tt)
	before calling setup().  This can support later joint use of
	multiple data sets allowing us to set the reference time for
	the second set from the peak time of the first.  The tmax
	parameter gives the shift versus this reference time.
  -mlsignal:
	-elimination of (meainingless) tstartHACK which had previously been
	available to recover the former even buggier version of tmax
	-eliminate the use of the older version of the clone_trajectory
	function which required explicit parameter information for parameters
	other than r0,phi, and required precomputing an offset time. The
	tmax scaling bug was discovered here.  In the current version
	we artificially rescale tmax in this part of the code to recover
	almost identical behavior to the previous version.  The results are
	indeed identical for typical evaluvations, and in a test run we
	obtain identical output for about 500 steps, but then differences
	appear.  The differences eventually grow to randomize all pt chains
	in my test run my 1000 steps.  It seems that the difference is
	caused by great sensitivity to tmax in rare cases where data points
	are very near the model caustic.  I have previously found that
	tiny changes in tmax (machine precision differences from the
	alternative calculation here) can lead to chain randomness because
	of such caustic issues.  Runnning the specific max posterior result
	from my old test run with the new code did yield identical results.
	-removed old comments and commented out some unused code.