No significant change, bookkeeping.
Running now with changes pulled into ptmcmc.  Not sure why, but there are tiny differences
in output precision and there are differences that quick change the evolution results.  For
now at least I assume that these are an unimportant consequence of the Feb 20 changes to
how extreme posteriors are handled.  More testing may be warranted.