Initial implementation of signal smearing.  If the new option -MLPsig_nsmear is set to something n<1 then when the magnification signal is computed it will be averaged over the n nearby points. The points are distributed in a regular, but approximately gaussian distributed pattern over some width set by another flag, with default: -MLPsig_dtsmear=0.002 given in units of Einstein time.
The idea is twofold:
	-This approximately mimics finite source effects with the
assumption that the source surface brightness is random in some
Gaussian area.  Part of the "approximately" is that we only smear
in 1-D along the trajectory direction.
	-If the finite source is a uniform disk, then the error we
make should be mitigated by the model variance here representing
our uncertainty about the surface brightness pattern.  Similarly
we should not be sensitive to surface features.
	-The other part of the idea is that this may help to improve
the Bayesian fitting.  Light curves with caustics are exceptionally
sensitive to the exact location of the caustic.  This means there are
extremely sharp variations in the liklihood over tiny changes in
parameter space.  This makes the optimization and characterization
much harder.  With this feature turned on the model can take into
account approximate information about the caustics in a potentially
beneficial way.
	-In the future I plan to make the width of the smearing region
a parameter.  This should allow the Bayesian search to begin with
fuzzy feautred signals and hone in as merited.  The ultimate best
fit regions for the smearing width may be a fair proxy for the finite
source width.