Change in default value for some trajectory parameters to match recent usage.
Added an automated test similar to what I have been using by hand.  This
test also reflects recent changes in the flags for ptmcmc.  An effort
has been made to reduce the number of flags, which maintaining
consistency with the test as it was before (though this size is
smaller).

To run the test:
cd test/m4-test
make