Fixed additional bug that was causing crashes in some long running finite_source runs.  The issue occurred with a paritcular combination of polynomial solve failures and polygon edge refinement.
I changed the logic a little so that polynomial failures should
actually trigger refinement, and fixed the original bug

There is still an issue with occasional messages about needing to
be "adding a point".  This is when the should be refinement at the
start/end of the loop around the polygon.  A draft fix for this is
coded, but not yet working, and currently disabled.

There is also some draft code for implementing brute force caluclation
of the finite source effect.  This is not yet tested or tied in. It
will be useful for comparisons to verify the area_mag code and it
will allow implementation of nontrivial source face imtensity profiles
which will be valuable in understanding the performance of the code
with the area mag variance implementation in these cases.