There were a sparse few points where the brute and polygon versions of the finite source code were not converging to the same result.  To resolve the difference I made another less brutish brute force code which essentially provides a coarse backward shooting method.  The new code actually starts with a coarse map of the lens contour using the polygon method, then defines a rectangular field of twice the dimensions needed for each of the resulting polygon contour regions.  These are gridded with point-resolution such that the pixel size is a fraction of the unlensed source area specified by finite_source_tol.  For each pixel the foward lens map to the source plane is computed and, as currently implemented, a simple test checks whether the point is inside the source disk or not.  If so then that pixels area is counted toward the magnification.
The current set up was used specifically for the mock-finite/test_run_e
cases t=5 (brute) and t=2 (polygon).  The polygon code is relatively
aggressively tuned for accuracy here, (though that seems to have minimal
effect on the outcome).  The result is that the brute methods now
converge toward the polygon method on the test problem (a zoomed-in
caustic crossing with 3 moderate to small finite source sizes). The
polygon code even with -GL_finite_source_Npoly_max=20 always does better
(usually by several orders) than the brute method with tol at 1e-4.

A future variation on the new brute method could be used to implement
non-trivial source-face intensity maps using the polygon contours as
a baseline.  The method could be made more efficient by estimating the
area ov overlap between the source-face and the pixel squares on the
edges.  A partially developed version of that is sketched in the code.